refactor, moved reasoning into own file
